Added support for request bodies and parameters
This is still very much a WIP, but it looks like it mostly builds
correctly (save for one more openapi3 bug I found).
